pppYmMana: CalcReflectionVector2 inline denom + pppFrameYmMana clean float store (#17518)
CalcReflectionVector2: inline the (denomBias +/- z) divisor into the -x/-y
divisions instead of a 'float denom' local, matching the target's eval order
(load x + negate before computing denom). misalign 9->5, byte-identical math.

pppFrameYmMana: retype pppYmManaStep::m_rippleLevel u32->float (pppYmMana.h is
included only by pppYmMana.cpp) and write heightA[144]=rippleLevel cleanly
instead of a reinterpret_cast hack; split the && so the pointer/value reads
schedule before the null guard (no null-deref introduced — heightA[144] stays
inside if(heightA!=0)). misalign 4->2. Supersedes #17513's reinterpret approach.

Verified with tools/instr_align.py (misalign drops, no other function regresses).

+++ include/ffcc/pppYmMana.h
@@ -32,7 +32,7 @@ struct pppYmManaStep {
     s32 m_envTextureId1;
     float m_waterScale;
     float m_waterOffset;
-    u32 m_rippleLevel;
+    float m_rippleLevel;
     u8 m_map21Flag;
     u8 m_pad39[3];
     GXColor m_baseColor;
